+++ R/obscureCounts.R
@@ -47,12 +47,15 @@ obscureCounts <- function(table,
       checkColNames <- c("n_records", "n_person")
     } else if (grepl("missingValues", tableName)) {
       checkColNames <- c("n_records", "n_person", "n_records_not_missing_value", "n_records_missing_value")
-    } else if (grepl("diagnosticsSummary|conceptSummary", tableName)) {
+    } else if (grepl("conceptSummary", tableName)) {
       checkColNames <- colNames <- c("n_records", "n_patients")
+    } else if (grepl("diagnosticsSummary", tableName)) {
+      checkColNames <- c("n_records", "n_patients")
+      colNames <- c("n_records", "n_patients")
     }
 
     # if any count col is less than minCellCount and larger than zero, replace colNames with substitute
-    if (!is.null(checkColNames)) {
+    if (!is.null(checkColNames) && nrow(table) > 0) {
       table <- table %>%
         dplyr::rowwise() %>%
         dplyr::mutate(result_obscured = any(dplyr::across(dplyr::all_of(checkColNames), ~ (. < minCellCount & . > 0)))) %>%