+++ R/summariseChecks.R
@@ -17,9 +17,11 @@
 #' Create a summary about the diagnostics results
 #'
 #' @param resultList a list with the diagnostics results
+#' @param minCellCount minimum number of events to report, results
+#' lower than this will be obscured. If 0 all results will be reported.
 #'
 #' @return a table containing the diagnostics summary
-summariseChecks <- function(resultList) {
+summariseChecks <- function(resultList, minCellCount = 5) {
   # total by ingredient
   dose_form_table <- resultList$conceptSummary %>%
     dplyr::group_by(.data$ingredient, .data$ingredient_concept_id) %>%
@@ -45,9 +47,13 @@ summariseChecks <- function(resultList) {
   # proportion with rxnorm dose form per ingredient
   diagnosticsSummary <- diagnosticsSummary %>%
     dplyr::mutate(n_dose_form = ifelse(is.na(.data$n_dose_form), 0, .data$n_dose_form)) %>%
+    dplyr::mutate(n_dose_form = ifelse(.data$n_dose_form > 0 & .data$n_dose_form < minCellCount, NA, .data$n_dose_form)) %>%
     dplyr::mutate(
       proportion_of_records_with_dose_form =
-        glue::glue("{.data$n_dose_form} ({round(100 * .data$n_dose_form / .data$n_records, 1)}%)")
+        dplyr::if_else(is.na(.data$n_dose_form),
+                       "NA (NA%)",
+                       glue::glue("{.data$n_dose_form} ({round(100 * .data$n_dose_form / .data$n_records, 1)}%)")
+        )
     ) %>%
     dplyr::select(-.data$n_dose_form)
 
@@ -66,6 +72,8 @@ summariseChecks <- function(resultList) {
           )) %>%
           dplyr::mutate(n_records_missing_value = dplyr::if_else(is.na(.data$n_records_missing_value), 0, .data$n_records_missing_value)) %>%
           dplyr::mutate(proportion_records_missing_value = dplyr::if_else(is.na(.data$proportion_records_missing_value), 0, .data$proportion_records_missing_value)) %>%
+          dplyr::mutate(n_records_missing_value = ifelse(.data$n_records_missing_value > 0 & .data$n_records_missing_value < minCellCount, NA, .data$n_records_missing_value)) %>%
+          dplyr::mutate(proportion_records_missing_value = ifelse(.data$n_records_missing_value > 0 & .data$n_records_missing_value < minCellCount, NA, .data$proportion_records_missing_value)) %>%
           dplyr::mutate(
             missing =
               glue::glue("
@@ -75,25 +83,25 @@ summariseChecks <- function(resultList) {
           tidyr::pivot_wider(names_from = .data$variable, values_from = .data$missing) %>%
           dplyr::mutate(
             missing_quantity_exp_start_end_days_supply =
-              glue::glue("
+              as.character(glue::glue("
         {.data$n_missing_quantity}, {.data$n_missing_drug_exposure_start_date}, {.data$n_missing_drug_exposure_end_date}, {.data$n_missing_days_supply}")
-          ) %>%
+          )) %>%
           dplyr::select("missing_quantity_exp_start_end_days_supply", "ingredient_concept_id"),
         by = "ingredient_concept_id"
       )
   } else {
     diagnosticsSummary <- diagnosticsSummary %>%
       dplyr::mutate(
-        missing_quantity_exp_start_end_days_supply =
-          glue::glue("0 (0%), 0 (0%), 0 (0%), 0 (0%)"))
+        missing_quantity_exp_start_end_days_supply = NA_character_)
   }
 
   # drug type
   if (!is.null(resultList$drugTypesOverall)) {
     totN <- as.numeric(sum(resultList$drugTypesOverall$n_records))
     drugTypes <- resultList$drugTypesOverall %>%
-      dplyr::mutate(drug_type_n = glue::glue("
-        {.data$drug_type} ({n_records}, {round(100 * .data$n_records / .env$totN, 1)}%)"))
+      dplyr::mutate(drug_type_n = dplyr::if_else(.data$n_records > 0 & .data$n_records < minCellCount,
+                                                 glue::glue("{.data$drug_type} (NA, NA%)"),
+                                                 glue::glue("{.data$drug_type} ({n_records}, {round(100 * .data$n_records / .env$totN, 1)}%)")))
     diagnosticsSummary$proportion_of_records_by_drug_type <- paste(drugTypes$drug_type_n, collapse = ";")
   }
 
@@ -101,8 +109,9 @@ summariseChecks <- function(resultList) {
   if (!is.null(resultList$drugRoutesOverall)) {
     totN <- as.numeric(sum(resultList$drugRoutesOverall$n_records))
     routes <- resultList$drugRoutesOverall %>%
-      dplyr::mutate(route_type_n = glue::glue("
-        {.data$route_type} ({n_records}, {round(100 * .data$n_records / .env$totN, 1)}%)"))
+      dplyr::mutate(route_type_n = dplyr::if_else(.data$n_records > 0 & .data$n_records < minCellCount,
+                                                  glue::glue("{.data$route_type} (NA, NA%)"),
+                                                  glue::glue("{.data$route_type} ({n_records}, {round(100 * .data$n_records / .env$totN, 1)}%)")))
     diagnosticsSummary$proportion_of_records_by_route_type <- paste(routes$route_type_n, collapse = ";")
   }
 
@@ -161,8 +170,8 @@ summariseChecks <- function(resultList) {
         ) %>% tidyr::pivot_wider(names_from = .data$estimate_name, values_from = .data$estimate_value) %>%
         dplyr::mutate(
           n_dose_and_missingness =
-            glue::glue(paste("{.data$count} ({.data$count_missing}, {.data$percentage_missing}%)"))
-        ) %>%
+            as.character(glue::glue(paste("{.data$count} ({.data$count_missing}, {.data$percentage_missing}%)"))
+        )) %>%
         dplyr::select("ingredient_concept_id", "n_dose_and_missingness"),
       by = "ingredient_concept_id"
     )
@@ -211,11 +220,16 @@ summariseChecks <- function(resultList) {
         ) %>%
         dplyr::mutate(
           median_daily_dose_q05_q95 =
-            glue::glue(paste("{.data$median} ({.data$q05}-{.data$q95}) [{.data$unit}]"))
-        ) %>%
+            as.character(glue::glue(paste("{.data$median} ({.data$q05}-{.data$q95}) [{.data$unit}]"))
+        )) %>%
         dplyr::select("ingredient_concept_id", "median_daily_dose_q05_q95"),
       by = "ingredient_concept_id"
     )
+  } else {
+    diagnosticsSummary <- diagnosticsSummary %>%
+      dplyr::mutate(
+        n_dose_and_missingness = NA_character_,
+        median_daily_dose_q05_q95 = NA_character_)
   }
 
   diagnosticsSummary <- diagnosticsSummary %>%