+++ tests/testthat/test-SyntheaSqlServer.R
@@ -10,14 +10,17 @@ test_that("test methods against test server", {
     Database = Sys.getenv("TESTDB_NAME"),
     UID      = Sys.getenv("TESTDB_USER"),
     PWD      = Sys.getenv("TESTDB_PWD"),
-    Port     = Sys.getenv("TESTDB_PORT")
+    Port     = Sys.getenv("TESTDB_PORT"),
+    bigint = c("numeric")
   )
-  cdm <- CDMConnector::cdm_from_con(con, cdm_schema = Sys.getenv("TESTDB_CDM_SCHEMA"))
+  cdm <- CDMConnector::cdmFromCon(con,
+                                  cdmSchema = Sys.getenv("TESTDB_CDM_SCHEMA"),
+                                  writeSchema = Sys.getenv("TESTDB_WRITE_SCHEMA"))
 
   result <- executeChecks(cdm = cdm, ingredients = c(1125315), verbose = TRUE) # acetaminophen
 
   # checks
-  expect_equal(length(result), 25)
+  expect_equal(length(result), 8)
   expect_true(all(grepl("acetaminophen", result$ingredientConcepts$concept_name)))
 
   DBI::dbDisconnect(attr(cdm, "dbcon"), shutdown = TRUE)